@@ -1,5 +1,5 @@
 {
-  "name": "Simserver-app",
+  "name": "simserver-app",
   "private": true,
   "version": "0.0.1",
   "type": "module",
@@ -10,21 +10,23 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^5.0.3",
-    "@tsconfig/svelte": "^5.0.2",
+    "@sveltejs/vite-plugin-svelte": "^6.0.0",
+    "@tailwindcss/postcss": "^4.1.11",
+    "@tsconfig/svelte": "^5.0.4",
     "autoprefixer": "^10.4.16",
-    "flowbite": "^2.3.0",
-    "flowbite-svelte": "^0.44.24",
-    "postcss": "^8.4.32",
-    "postcss-load-config": "^5.0.2",
-    "svelte": "^4.2.12",
-    "svelte-check": "^3.6.7",
-    "tailwindcss": "^3.3.6",
-    "tslib": "^2.6.2",
-    "typescript": "^5.2.2",
-    "vite": "^6.3.4"
+    "postcss": "^8.5.6",
+    "postcss-load-config": "^6.0.1",
+    "svelte": "^5.36.0",
+    "svelte-check": "^4.3.0",
+    "tailwindcss": "^4.1.11",
+    "tslib": "^2.8.1",
+    "typescript": "^5.8.3",
+    "vite": "^7.0.0"
   },
   "dependencies": {
-    "axios": "^1.8.2"
+    "@tailwindcss/vite": "^4.1.11",
+    "axios": "^1.7.7",
+    "clsx": "^2.0.0",
+    "tailwind-merge": "^2.0.0"
   }
 }
