@@ -1,12 +1,8 @@
-const tailwindcss = require("tailwindcss");
-const autoprefixer = require("autoprefixer");
-
 const config = {
-  plugins: [//Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-  //But others, like autoprefixer, need to run after,
-  tailwindcss(),
-] //Some plugins, like tailwindcss/nesting, need to run before Tailwind, tailwindcss(), //But others, like autoprefixer, need to run after, autoprefixer, autoprefixer],
-
+  plugins: [
+    require('@tailwindcss/postcss'),
+    require('autoprefixer'),
+  ],
 };
 
 module.exports = config;
