@@ -1,7 +1,17 @@
 import { defineConfig } from 'vite'
 import { svelte } from '@sveltejs/vite-plugin-svelte'
+import tailwindcss from '@tailwindcss/vite'
+import { resolve } from 'path'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [svelte()],
+  plugins: [
+    svelte(),
+    tailwindcss()
+  ],
+  resolve: {
+    alias: {
+      $lib: resolve('./src/lib')
+    }
+  }
 })
